fix: 011 建议+可选修复——M3/M5/M6/L2/L3/L4/L6
M3: ownerId 参数统一为 Long(selectPageByOwnerId 从 Integer 改 Long)
M5: @Transactional 删冗余 rollbackFor,补事务边界 Javadoc
M6: division @RequestParam(required=false),缺失走统一中文错误
L2: selectPageByOwnerId Javadoc 修正为所有者ID
L3: division 补 @Schema 注解
L4: syncCharacterJournal 显式回填 ownerId
L6: RBAC 脚本补 DELETE clobbers 警告 + 预检 SELECT

M1(VALUES 弃用)技术栈冻结阻断:Druid 1.2.24 无法解析行别名语法,回退并注释,待解冻
测试 24/24 通过

+++ src/main/java/xyz/foolcat/eve/evehelper/application/service/WalletJournalApplicationService.java
@@ -99,7 +99,7 @@ public class WalletJournalApplicationService {
         // 归属校验先于业务逻辑:cid 为用户可控入参,须先确认该人物属于当前用户(防御 IDOR)
         accessGuard.requireOwnership(String.valueOf(cid), "钱包流水");
         IPage<WalletJournal> page = new Page<>(current, size);
-        IPage<WalletJournal> domainPage = walletJournalRepository.selectPageByOwnerId(page, cid);
+        IPage<WalletJournal> domainPage = walletJournalRepository.selectPageByOwnerId(page, cid.longValue());
         return PageResultUtil.copy(domainPage, walletJournalAssembler::toVo);
     }
